requirements.txt
pip~=23.1.2
wheel~=0.40.0
setuptools~=67.8.0
asgiref==3.7.2
Django==4.2.3
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
Pillow==10.0.0
psycopg2-binary==2.9.6
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
drf_yasg==1.21.7